Add infotooltip to details menu items
